requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
azure-common==1.1.28
azure-core==1.32.0
azure-cosmos==4.9.0
azure-search-documents==11.5.2
azure-storage-blob==12.24.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
databricks==0.2
databricks-sql-connector==2.9.3
DateTime==5.5
distro==1.9.0
et_xmlfile==2.0.0
Flask==3.1.0
Flask-Cors==5.0.0
fuzzywuzzy==0.18.0
google-auth==2.36.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.1
Levenshtein==0.26.1
lz4==4.3.3
Mako==1.3.6
MarkupSafe==3.0.2
numpy==2.1.3
oauthlib==3.2.2
openai==1.54.5
openpyxl==3.1.5
pandas==2.2.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-Levenshtein==0.26.1
pytz==2024.2
RapidFuzz==3.10.1
requests==2.32.3
rsa==4.9
setuptools==75.5.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.54
thrift==0.16.0
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
wheel==0.45.1
zope.interface==7.1.1
